+++ src/main/java/com/unimelbCoder/melbcode/models/dao/CommentDao.java
@@ -9,9 +9,9 @@ import java.util.List;
 @Repository
 public interface CommentDao {
 
-    public Comment getCommentById(@Param("id") Integer id);
+    public Comment getCommentById(@Param("id") Long id);
 
-    public Comment getCommentByIdxArticle(@Param("article_id") Integer article_id);
+    public Comment getCommentByIdxArticle(@Param("article_id") Long article_id);
 
     public Comment getCommentByIdxUserId(@Param("user_id") Integer user_id);
 
@@ -20,22 +20,22 @@ public interface CommentDao {
      * @param article_id
      * @return
      */
-    public List<Comment> getTopCommentList(@Param("article_id") Integer article_id);
+    public List<Comment> getTopCommentList(@Param("article_id") Long article_id);
 
-    public List<Comment> getSubCommentList(@Param("article_id") Integer article_id,
-                                           @Param("top_comment_id") Integer top_comment_id);
+    public List<Comment> getSubCommentList(@Param("article_id") Long article_id,
+                                           @Param("top_comment_id") Long top_comment_id);
 
     /**
      * 查看文章有效评论数
      * @param article_id
      * @return
      */
-    public Integer getValidCommentCount(@Param("article_id") Integer article_id);
+    public Integer getValidCommentCount(@Param("article_id") Long article_id);
 
-    public void createComment(@Param("article_id") Integer article_id,
+    public void createComment(@Param("article_id") Long article_id,
                               @Param("user_id") String user_id,
                               @Param("content") String content,
-                              @Param("top_comment_id") Integer top_comment_id,
-                              @Param("parent_comment_id") Integer parent_comment_id);
+                              @Param("top_comment_id") Long top_comment_id,
+                              @Param("parent_comment_id") Long parent_comment_id);
 
 }
